% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api.R
\name{api_get}
\alias{api_get}
\title{GET request helper}
\usage{
api_get(config, endpoint, params = list())
}
\arguments{
\item{config}{API configuration}

\item{endpoint}{API endpoint}

\item{params}{Query parameters}
}
\value{
API response
}
\description{
GET request helper
}
