% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_api.R
\name{api_post}
\alias{api_post}
\title{POST request helper}
\usage{
api_post(config, endpoint, body, params = list())
}
\arguments{
\item{config}{API configuration}

\item{endpoint}{API endpoint}

\item{body}{Request body}

\item{params}{Query parameters}
}
\value{
API response
}
\description{
POST request helper
}
