% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{cancel_slurm_job}
\alias{cancel_slurm_job}
\title{Cancel Slurm Job}
\usage{
cancel_slurm_job(slurm_job_id)
}
\arguments{
\item{slurm_job_id}{Character string with the Slurm job ID.}
}
\value{
Logical indicating whether the cancellation was successful.
}
\description{
Cancels a running or pending Slurm job.
}
\keyword{internal}
