% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hashing.R
\name{check_existing_hashes}
\alias{check_existing_hashes}
\title{Check which hashes already exist in the database}
\usage{
check_existing_hashes(config, hashes)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{hashes}{Character vector of file hashes to check}
}
\value{
A list with two components: existing_hashes and missing_hashes
}
\description{
Check which hashes already exist in the database
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
hash1 <- hash_file("file1.txt")
hash2 <- hash_file("file2.txt")
result <- check_existing_hashes(config, c(hash1, hash2))
}
}
