% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{create_job_script}
\alias{create_job_script}
\title{Create Job Script}
\usage{
create_job_script(
  func_name,
  args,
  job_id,
  db_path,
  required_packages = NULL,
  slurm_opts = list()
)
}
\arguments{
\item{func_name}{Character string with the name of the function.}

\item{args}{List of arguments to pass to the function.}

\item{job_id}{Character string with the job ID.}

\item{db_path}{Character string with the path to the database.}

\item{required_packages}{Character vector with additional packages to load.}

\item{slurm_opts}{List with Slurm-specific options.}
}
\value{
Character string with the job script.
}
\description{
Creates an R script for running a job.
}
\keyword{internal}
