% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{create_pipeline_node}
\alias{create_pipeline_node}
\title{Create a pipeline node definition}
\usage{
create_pipeline_node(
  chain,
  dependencies = character(0),
  input_file_hash = NULL
)
}
\arguments{
\item{chain}{List of method steps (each can have file_inputs)}

\item{dependencies}{Vector of node IDs this depends on}

\item{input_file_hash}{Optional file hash for root nodes}
}
\value{
Node definition
}
\description{
Create a pipeline node definition
}
\details{
Each step in the chain can now include file_inputs with $ref support:
\code{
chain = list(list(
  method_name = "merge_files",
  parameters = list(separator = "|"),
  file_inputs = list(
    input_1 = "$ref:node1/data/text",
    input_2 = "$ref:node2/data/text"
  )
))
}
}
