% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{create_python_script}
\alias{create_python_script}
\title{Create Python Script for Job Execution}
\usage{
create_python_script(
  py_module,
  py_function,
  args,
  result_file,
  python_path = NULL,
  virtualenv = NULL,
  condaenv = NULL,
  required_modules = NULL
)
}
\arguments{
\item{py_module}{Character string with the name of the Python module.}

\item{py_function}{Character string with the name of the Python function.}

\item{args}{List of arguments to pass to the Python function.}

\item{result_file}{Character string with the path where the result should be saved.}

\item{python_path}{Optional character string with the path to the Python executable.}

\item{virtualenv}{Optional character string with the name of the virtualenv to use.}

\item{condaenv}{Optional character string with the name of the conda environment to use.}

\item{required_modules}{Character vector with additional Python modules to import.}
}
\value{
Path to the created Python script.
}
\description{
Creates a Python script for executing a function from a module.
}
\keyword{internal}
