% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{create_python_wrapper}
\alias{create_python_wrapper}
\title{Create Python Wrapper Function}
\usage{
create_python_wrapper(
  py_module,
  py_function,
  args = list(),
  python_path = NULL,
  virtualenv = NULL,
  condaenv = NULL,
  required_modules = NULL
)
}
\arguments{
\item{py_module}{Character string with the name of the Python module.}

\item{py_function}{Character string with the name of the Python function.}

\item{args}{List of arguments to pass to the Python function.}

\item{python_path}{Optional character string with the path to the Python executable.}

\item{virtualenv}{Optional character string with the name of the virtualenv to use.}

\item{condaenv}{Optional character string with the name of the conda environment to use.}

\item{required_modules}{Character vector with additional Python modules to import.}
}
\value{
An R function that will execute the Python function.
}
\description{
Creates an R function that will execute a Python function via reticulate.
}
\keyword{internal}
