% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{create_r_script}
\alias{create_r_script}
\title{Create R Execution Script}
\usage{
create_r_script(
  function_name,
  args,
  package = NULL,
  result_file,
  required_packages = NULL
)
}
\arguments{
\item{function_name}{Character string with the name of the function to execute.}

\item{args}{List of arguments to pass to the function.}

\item{package}{Character string with the package name containing the function.}

\item{result_file}{Character string with the path where the result should be saved.}

\item{required_packages}{Character vector with names of packages to load.}
}
\value{
Path to the created R script.
}
\description{
Creates a temporary R script that will execute the specified function with arguments.
}
\keyword{internal}
