% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{.backend_kubernetes_contract}
\alias{.backend_kubernetes_contract}
\title{Kubernetes backend contract resolved from server options}
\usage{
.backend_kubernetes_contract(settings = .dshpc_settings())
}
\description{
Kubernetes backend contract resolved from server options
}
\keyword{internal}
