% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical.R
\name{.canonicalise_spec}
\alias{.canonicalise_spec}
\title{Canonicalise a job spec object before JSON hashing}
\usage{
.canonicalise_spec(x)
}
\description{
Canonicalise a job spec object before JSON hashing
}
\keyword{internal}
