% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.dshpc_home_health}
\alias{.dshpc_home_health}
\title{Check DSHPC_HOME directories and permissions.}
\usage{
.dshpc_home_health()
}
\description{
Check DSHPC_HOME directories and permissions.
}
\keyword{internal}
