% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.dsjobs_home_health}
\alias{.dsjobs_home_health}
\title{Check DSJOBS_HOME directories and permissions.}
\usage{
.dsjobs_home_health()
}
\description{
Check DSJOBS_HOME directories and permissions.
}
\keyword{internal}
