% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.dsjobs_worker_start}
\alias{.dsjobs_worker_start}
\title{Start the worker daemon (admin/setup helper, NOT auto-start)}
\usage{
.dsjobs_worker_start()
}
\description{
Start the worker daemon (admin/setup helper, NOT auto-start)
}
\keyword{internal}
