% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executor.R
\name{.executor_run_step}
\alias{.executor_run_step}
\title{Execute the current step of a job}
\usage{
.executor_run_step(db, job_id, step_index, spec)
}
\description{
Execute the current step of a job
}
\keyword{internal}
