% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.generate_job_id}
\alias{.generate_job_id}
\title{Generate unique job ID (UUIDv4, 122 bits entropy)}
\usage{
.generate_job_id()
}
\description{
Generate unique job ID (UUIDv4, 122 bits entropy)
}
\keyword{internal}
