% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.leader_process_alive}
\alias{.leader_process_alive}
\title{Check whether a scheduler leader still has a live process}
\usage{
.leader_process_alive(db, leader)
}
\description{
Check whether a scheduler leader still has a live process
}
\keyword{internal}
