% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plugins.R
\name{.load_publisher_packages}
\alias{.load_publisher_packages}
\title{Load publisher plugin packages requested by a job step}
\usage{
.load_publisher_packages(step)
}
\description{
Load publisher plugin packages requested by a job step
}
\keyword{internal}
