% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.mark_workers_stopped}
\alias{.mark_workers_stopped}
\title{Mark stopped workers in the scheduler catalog}
\usage{
.mark_workers_stopped(pids = integer(0), worker_ids = character(0))
}
\description{
Mark stopped workers in the scheduler catalog
}
\keyword{internal}
