% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{.normalize_output_size_bytes}
\alias{.normalize_output_size_bytes}
\title{Register an output in the outputs table}
\usage{
.normalize_output_size_bytes(size_bytes)
}
\description{
Register an output in the outputs table
}
\keyword{internal}
