% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package attachment}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{The library name}

\item{pkgname}{The package name}
}
\description{
Package attachment
}
\keyword{internal}
