% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package Initialization Functions}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{The library name}

\item{pkgname}{The package name}
}
\description{
Functions that run when the package is loaded or attached.
Package initialization
}
\keyword{internal}
