% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.output_object_cardinality}
\alias{.output_object_cardinality}
\title{Establish the row cardinality of a deserialized output}
\usage{
.output_object_cardinality(obj)
}
\value{
Integer cardinality, or NA when the object has no unambiguous rows.
}
\description{
Establish the row cardinality of a deserialized output
}
\keyword{internal}
