% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner_artifact.R
\name{.proc_get_exit}
\alias{.proc_get_exit}
\title{Get exit status from processx handle, clean up registry}
\usage{
.proc_get_exit(job_id, step_index)
}
\description{
Get exit status from processx handle, clean up registry
}
\keyword{internal}
