% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner_artifact.R
\name{.proc_is_alive}
\alias{.proc_is_alive}
\title{Check if a job's artifact step is still running via processx handle}
\usage{
.proc_is_alive(job_id, step_index)
}
\description{
Check if a job's artifact step is still running via processx handle
}
\keyword{internal}
