% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{.require_job_access}
\alias{.require_job_access}
\title{Require possession of a job bearer capability}
\usage{
.require_job_access(db, job_id_or_symbol)
}
\description{
Require possession of a job bearer capability
}
\keyword{internal}
