% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.worker_runtime_snapshot}
\alias{.worker_runtime_snapshot}
\title{Runtime snapshot of known worker processes}
\usage{
.worker_runtime_snapshot(home = .dshpc_home())
}
\description{
Runtime snapshot of known worker processes
}
\keyword{internal}
