% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.worker_write_health}
\alias{.worker_write_health}
\title{Write worker health file (for monitoring)}
\usage{
.worker_write_health()
}
\description{
Write worker health file (for monitoring)
}
\keyword{internal}
