% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHPC_methods.R
\name{dsHPC.ds.cancel}
\alias{dsHPC.ds.cancel}
\title{Cancel a job in DataSHIELD environment}
\usage{
dsHPC.ds.cancel(job_id)
}
\arguments{
\item{job_id}{Character string with the job ID.}
}
\value{
Logical indicating whether the cancellation was successful.
}
\description{
DataSHIELD-compatible version of dsHPC.cancel.
}
