% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHPC_methods.R
\name{dsHPC.ds.list_jobs}
\alias{dsHPC.ds.list_jobs}
\title{List jobs in DataSHIELD environment}
\usage{
dsHPC.ds.list_jobs(status_filter = NULL)
}
\arguments{
\item{status_filter}{Optional character vector with statuses to filter by.}
}
\value{
A data frame containing job information.
}
\description{
DataSHIELD-compatible version of dsHPC.list_jobs.
}
