% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHPC_methods.R
\name{dsHPC.ds.status}
\alias{dsHPC.ds.status}
\title{Get job status in DataSHIELD environment}
\usage{
dsHPC.ds.status(job_id)
}
\arguments{
\item{job_id}{Character string with the job ID.}
}
\value{
A list containing job status information.
}
\description{
DataSHIELD-compatible version of dsHPC.status.
}
