% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHPC_methods.R
\name{dsHPC.ds.submit}
\alias{dsHPC.ds.submit}
\title{Submit a function execution as a job in DataSHIELD environment}
\usage{
dsHPC.ds.submit(
  func_name,
  args,
  package = NULL,
  required_packages = NULL,
  object_hash = NULL,
  use_cache = TRUE,
  slurm_opts = list()
)
}
\arguments{
\item{func_name}{Character string with the name of the function to execute.}

\item{args}{List of arguments to pass to the function.}

\item{package}{Character string with the package name containing the function.}

\item{required_packages}{Character vector with names of additional packages to load.}

\item{object_hash}{Optional character string with a hash identifying a data object.}

\item{use_cache}{Logical indicating whether to check for cached results.}

\item{slurm_opts}{List of options for Slurm job submission.}
}
\value{
A list containing job information.
}
\description{
DataSHIELD-compatible version of dsHPC.submit.
}
