% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsHPC_methods.R
\name{dsHPC.wrap_function}
\alias{dsHPC.wrap_function}
\title{DataSHIELD Function Wrapping Utility}
\usage{
dsHPC.wrap_function(
  func,
  package = NULL,
  required_packages = NULL,
  slurm_opts = list(),
  default_wait = TRUE
)
}
\arguments{
\item{func}{Function to wrap.}

\item{package}{Character string with the package name containing the function.}

\item{required_packages}{Character vector with names of additional packages to load.}

\item{slurm_opts}{List of options for Slurm job submission.}

\item{default_wait}{Logical indicating whether to wait for completion by default.}
}
\value{
A function that will execute the original function via HPC.
}
\description{
Creates a DataSHIELD-compatible wrapped version of a function for execution in HPC.
}
