% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_resource.R
\name{DsHpcUnitResourceClient}
\alias{DsHpcUnitResourceClient}
\title{dsHPC execution unit Resource client}
\description{
dsHPC execution unit Resource client

dsHPC execution unit Resource client
}
\details{
Resolves a `dshpc+unit://` Resource to a sealed, non-secret execution-unit
snapshot from the server administrator catalogue. Generic data-frame
materialization is deliberately unavailable.
}
\section{Super class}{
\code{\link[resourcer:ResourceClient]{resourcer::ResourceClient}} -> \code{DsHpcUnitResourceClient}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DsHpcUnitResourceClient-new}{\code{DsHpcUnitResourceClient$new()}}
\item \href{#method-DsHpcUnitResourceClient-getUnitSelection}{\code{DsHpcUnitResourceClient$getUnitSelection()}}
\item \href{#method-DsHpcUnitResourceClient-asDataFrame}{\code{DsHpcUnitResourceClient$asDataFrame()}}
\item \href{#method-DsHpcUnitResourceClient-clone}{\code{DsHpcUnitResourceClient$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="asTbl"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-asTbl'><code>resourcer::ResourceClient$asTbl()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="close"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-close'><code>resourcer::ResourceClient$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="downloadFile"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-downloadFile'><code>resourcer::ResourceClient$downloadFile()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="exec"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-exec'><code>resourcer::ResourceClient$exec()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getConnection"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getConnection'><code>resourcer::ResourceClient$getConnection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceClient" data-id="getResource"><a href='../../resourcer/html/ResourceClient.html#method-ResourceClient-getResource'><code>resourcer::ResourceClient$getResource()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceClient-new"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceClient-new}{}}}
\subsection{Method \code{new()}}{
Create a client for an assigned execution-unit Resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceClient$new(resource)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resource}}{A DataSHIELD Resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceClient-getUnitSelection"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceClient-getUnitSelection}{}}}
\subsection{Method \code{getUnitSelection()}}{
Return the validated, non-secret unit snapshot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceClient$getUnitSelection()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceClient-asDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceClient-asDataFrame}{}}}
\subsection{Method \code{asDataFrame()}}{
Refuse generic materialization of a control Resource.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceClient$asDataFrame()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceClient-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceClient-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceClient$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
