% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_resource.R
\name{DsHpcUnitResourceResolver}
\alias{DsHpcUnitResourceResolver}
\title{dsHPC execution unit Resource resolver}
\description{
dsHPC execution unit Resource resolver

dsHPC execution unit Resource resolver
}
\section{Super class}{
\code{\link[resourcer:ResourceResolver]{resourcer::ResourceResolver}} -> \code{DsHpcUnitResourceResolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DsHpcUnitResourceResolver-isFor}{\code{DsHpcUnitResourceResolver$isFor()}}
\item \href{#method-DsHpcUnitResourceResolver-newClient}{\code{DsHpcUnitResourceResolver$newClient()}}
\item \href{#method-DsHpcUnitResourceResolver-clone}{\code{DsHpcUnitResourceResolver$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="resourcer" data-topic="ResourceResolver" data-id="initialize"><a href='../../resourcer/html/ResourceResolver.html#method-ResourceResolver-initialize'><code>resourcer::ResourceResolver$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceResolver-isFor"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceResolver-isFor}{}}}
\subsection{Method \code{isFor()}}{
Test whether a Resource uses the dsHPC unit URL scheme.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceResolver$isFor(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A DataSHIELD Resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceResolver-newClient"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceResolver-newClient}{}}}
\subsection{Method \code{newClient()}}{
Create the dsHPC Resource client.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceResolver$newClient(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A DataSHIELD Resource descriptor.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DsHpcUnitResourceResolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DsHpcUnitResourceResolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DsHpcUnitResourceResolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
