% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{execute_pipeline}
\alias{execute_pipeline}
\title{Execute pipeline and wait for results}
\usage{
execute_pipeline(
  config,
  pipeline_definition,
  timeout = NA,
  interval = 5,
  verbose = TRUE
)
}
\arguments{
\item{config}{API configuration}

\item{pipeline_definition}{Pipeline definition}

\item{timeout}{Maximum time to wait (default: NA)}

\item{interval}{Polling interval (default: 5)}

\item{verbose}{Show progress (default: TRUE)}
}
\value{
Final output from pipeline
}
\description{
Execute pipeline and wait for results
}
\examples{
\dontrun{
pipeline <- list(
  nodes = list(
    node_1 = list(
      chain = list(list(method_name = "concat", parameters = list(a = "A", b = "B"))),
      dependencies = list()
    ),
    node_2 = list(
      chain = list(list(method_name = "concat", parameters = list(a = "C", b = "D"))),
      dependencies = list()
    ),
    node_3 = list(
      chain = list(list(method_name = "concat", parameters = list(a = "$ref:node_1", b = "$ref:node_2"))),
      dependencies = list("node_1", "node_2")
    )
  )
)

result <- execute_pipeline(config, pipeline)
}
}
