% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hashing.R
\name{file_exists}
\alias{file_exists}
\title{Check if a file already exists in the database based on its hash}
\usage{
file_exists(config, file_path)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{file_path}{Path to the file to check}
}
\value{
Boolean indicating if the file already exists in the database
}
\description{
Check if a file already exists in the database based on its hash
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
exists <- file_exists(config, "data.csv")
if (exists) {
  print("File already exists")
} else {
  print("File needs to be uploaded")
}
}
}
