% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{find_jobs_by_hash}
\alias{find_jobs_by_hash}
\title{Find Jobs by Hash}
\usage{
find_jobs_by_hash(con, job_hash)
}
\arguments{
\item{con}{DBI connection object to the SQLite database.}

\item{job_hash}{Character string with the job hash.}
}
\value{
A data frame with matching jobs or empty data frame if none found.
}
\description{
Finds jobs with a specific hash value.
}
\keyword{internal}
