% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_resource.R
\name{get_api}
\alias{get_api}
\title{Get an API client from a resource}
\usage{
get_api(resource)
}
\arguments{
\item{resource}{A list containing resource configuration (url, format, secret)}
}
\value{
An HPCResourceClient instance
}
\description{
Get an API client from a resource
}
\examples{
\dontrun{
resource <- list(
  url = "http://localhost:9000",
  format = "dshpc.api",
  secret = "please_change_me"
)

api <- get_api(resource)
}
}
