% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_cached_result}
\alias{get_cached_result}
\title{Retrieve Job Result}
\usage{
get_cached_result(con, job_hash)
}
\arguments{
\item{con}{DBI connection object to the SQLite database.}

\item{job_hash}{Character string with the hash of the job.}
}
\value{
The cached R object if found, or NULL if not cached.
}
\description{
Retrieves a cached job result from the database.
}
\keyword{internal}
