% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{get_job_info}
\alias{get_job_info}
\title{Get Job Information}
\usage{
get_job_info(con, job_id)
}
\arguments{
\item{con}{DBI connection object to the SQLite database.}

\item{job_id}{Character string with the job ID.}
}
\value{
A data frame with job information or NULL if not found.
}
\description{
Retrieves information about a specific job.
}
\keyword{internal}
