% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_jobs.R
\name{get_job_output_by_hash}
\alias{get_job_output_by_hash}
\title{Get the output of a completed job by hash}
\usage{
get_job_output_by_hash(
  config,
  file_hash,
  method_name,
  parameters = list(),
  parse_json = TRUE
)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{file_hash}{Hash of the content (as returned by hash_content)}

\item{method_name}{Name of the method executed}

\item{parameters}{Parameters used for the method}

\item{parse_json}{Whether to parse the output as JSON (default: TRUE)}
}
\value{
The job output, parsed as JSON if requested
}
\description{
Get the output of a completed job by hash
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
file_hash <- "abc123..."
output <- get_job_output_by_hash(config, file_hash, "count_black_pixels", list(threshold = 30))
}
}
