% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_jobs.R
\name{get_job_status}
\alias{get_job_status}
\title{Get the status of a job}
\usage{
get_job_status(config, file_path, method_name, parameters = list())
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{file_path}{Path to the file processed}

\item{method_name}{Name of the method executed}

\item{parameters}{Parameters used for the method}
}
\value{
The status of the job as a string
}
\description{
Get the status of a job
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
status <- get_job_status(config, "image.jpg", "count_black_pixels", list(threshold = 30))
}
}
