% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_meta.R
\name{get_meta_job_cache_info}
\alias{get_meta_job_cache_info}
\title{Get cached steps from a meta-job}
\usage{
get_meta_job_cache_info(config, meta_job_hash)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{meta_job_hash}{ID of the meta-job to check}
}
\value{
A data frame with step information including cache usage
}
\description{
Returns information about which steps in a meta-job used cached results
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
cache_info <- get_meta_job_cache_info(config, "meta-job-uuid-here")
}
}
