% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_meta.R
\name{get_meta_job_status}
\alias{get_meta_job_status}
\title{Get the status of a meta-job}
\usage{
get_meta_job_status(config, meta_job_id)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{meta_job_id}{ID of the meta-job to check}
}
\value{
A list with detailed meta-job status including chain progress
}
\description{
Get the status of a meta-job
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
status <- get_meta_job_status(config, "meta-job-uuid-here")
}
}
