% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{get_pipeline_status}
\alias{get_pipeline_status}
\title{Get pipeline status}
\usage{
get_pipeline_status(config, pipeline_hash)
}
\arguments{
\item{config}{API configuration}

\item{pipeline_hash}{Pipeline ID}
}
\value{
Pipeline status information
}
\description{
Get pipeline status
}
