% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hashing.R
\name{hash_content}
\alias{hash_content}
\title{Generate a hash based on content}
\usage{
hash_content(content)
}
\arguments{
\item{content}{The content to hash (raw vector, character, or other object)}
}
\value{
A character string with the content hash using SHA-256 algorithm
}
\description{
Generate a hash based on content
}
\examples{
\dontrun{
content <- "Hello World"
content_hash <- hash_content(content)
}
}
