% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_hashing.R
\name{hash_exists}
\alias{hash_exists}
\title{Check if a specific hash exists in the database}
\usage{
hash_exists(config, hash)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{hash}{A single hash string to check}
}
\value{
Boolean indicating if the hash exists
}
\description{
Check if a specific hash exists in the database
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
content <- "Hello World"
hash <- hash_content(content)
if (hash_exists(config, hash)) {
  print("Hash exists in database")
}
}
}
