% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking.R
\name{hpcTrackingStatusInternal}
\alias{hpcTrackingStatusInternal}
\title{Read Tracking Status from Trusted Server Code}
\usage{
hpcTrackingStatusInternal(tracking_id)
}
\arguments{
\item{tracking_id}{Tracking root id.}
}
\value{
Disclosure-safe coarse status.
}
\description{
Read Tracking Status from Trusted Server Code
}
