% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{init_db_connection}
\alias{init_db_connection}
\title{Database Management Functions for dsHPC}
\usage{
init_db_connection(db_path = NULL)
}
\arguments{
\item{db_path}{Character string specifying the path to the SQLite database file.
If NULL (default), uses the package's default location.}
}
\value{
A DBI connection object to the database.
}
\description{
This file contains functions for interacting with the SQLite database
that stores job information and results for caching purposes.

Creates and initializes the SQLite database for storing job information.
}
\keyword{internal}
