% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_datashield_env}
\alias{is_datashield_env}
\title{Check if Running in DataSHIELD Environment}
\usage{
is_datashield_env()
}
\value{
Logical indicating whether running in a DataSHIELD environment.
}
\description{
Determines if the code is being executed in a DataSHIELD environment.
}
\keyword{internal}
