% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{is_result_cached}
\alias{is_result_cached}
\title{Check if Job Result is Cached}
\usage{
is_result_cached(con, job_hash)
}
\arguments{
\item{con}{DBI connection object to the SQLite database.}

\item{job_hash}{Character string with the job hash to check.}
}
\value{
Logical indicating whether the job result is cached.
}
\description{
Checks if a result for a specific job hash exists in the cache.
}
\keyword{internal}
