% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm.R
\name{is_slurm_available}
\alias{is_slurm_available}
\title{Slurm Job Scheduler Interface for dsHPC}
\usage{
is_slurm_available()
}
\value{
Logical indicating whether Slurm is available.
}
\description{
This file contains functions for interacting with the Slurm job scheduler.

Checks if Slurm commands are available on the system.
}
\keyword{internal}
