% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{jobAdminCancelDS}
\alias{jobAdminCancelDS}
\title{Cancel Any Job (admin only)}
\usage{
jobAdminCancelDS(job_id, admin_key = NULL)
}
\arguments{
\item{job_id}{Character; job ID.}

\item{admin_key}{Character; the admin key.}
}
\description{
Disabled by default. Enable by setting dsjobs.admin_key or DSJOBS_ADMIN_KEY.
}
