% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{jobListDS}
\alias{jobListDS}
\title{List Jobs}
\usage{
jobListDS(label = NULL)
}
\arguments{
\item{label}{Character or NULL; filter by label.}
}
\description{
Returns all jobs, optionally filtered by label.
}
