% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{store_job_result}
\alias{store_job_result}
\title{Store Job Result}
\usage{
store_job_result(con, job_hash, result)
}
\arguments{
\item{con}{DBI connection object to the SQLite database.}

\item{job_hash}{Character string with the hash of the job.}

\item{result}{The R object containing the result.}
}
\value{
Logical indicating success.
}
\description{
Stores a job result in the cache.
}
\keyword{internal}
