% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_multi_file.R
\name{submit_meta_job_multi}
\alias{submit_meta_job_multi}
\title{Submit a meta-job with multiple input files}
\usage{
submit_meta_job_multi(config, file_inputs, method_chain)
}
\arguments{
\item{config}{API configuration}

\item{file_inputs}{Named list of file hashes (e.g., list(primary="hash1", reference="hash2"))}

\item{method_chain}{List of method specifications}
}
\value{
Response from API with meta_job_hash
}
\description{
Submit a meta-job with multiple input files
}
