% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{submit_pipeline}
\alias{submit_pipeline}
\title{Submit a pipeline (DAG of meta-jobs)}
\usage{
submit_pipeline(config, pipeline_definition)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{pipeline_definition}{List with 'nodes'}
}
\value{
Pipeline response with pipeline_hash
}
\description{
Submit a pipeline (DAG of meta-jobs)
}
\examples{
\dontrun{
pipeline <- list(
  nodes = list(
    node_1 = list(
      chain = list(list(method_name = "concat", parameters = list(a = "Hello", b = "World"))),
      dependencies = list()
    ),
    node_2 = list(
      chain = list(list(method_name = "concat", parameters = list(a = "$ref:node_1", b = "!"))),
      dependencies = list("node_1")
    )
  )
)

response <- submit_pipeline(config, pipeline)
}
}
