% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python.R
\name{to_python_arg}
\alias{to_python_arg}
\title{Convert R Object to Python-compatible String Representation}
\usage{
to_python_arg(arg)
}
\arguments{
\item{arg}{The R object to convert.}
}
\value{
A character string with the Python representation of the object.
}
\description{
Converts an R object to a string that can be used in a Python script.
}
\keyword{internal}
