% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_compress.R
\name{upload_file_compressed}
\alias{upload_file_compressed}
\title{Upload compressed content to the HPC resource}
\usage{
upload_file_compressed(
  config,
  content,
  filename,
  compression_type = c("gzip", "xz", "bzip2"),
  compress = TRUE
)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{content}{Content to compress and upload (raw vector, character, or other object)}

\item{filename}{Name to use for the uploaded content (will have compression extension added)}

\item{compression_type}{Compression type: "gzip" (default), "xz", or "bzip2"}

\item{compress}{Whether to compress the content (default TRUE)}
}
\value{
TRUE if the content was successfully uploaded or already exists
}
\description{
Upload compressed content to the HPC resource
}
\details{
This function compresses content deterministically before uploading, which means:
\itemize{
\item Same content will always produce the same hash
\item Content is deduplicated effectively in the cache
\item Reduced bandwidth usage
\item The server will need to decompress the content when needed
}
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
content <- "Hello, World!"
upload_file_compressed(config, content, "hello.txt", compression_type = "gzip")
}
}
