% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_pipeline.R
\name{create_pipeline_node}
\alias{create_pipeline_node}
\title{Create a pipeline node definition}
\usage{
create_pipeline_node(
  chain,
  dependencies = character(0),
  input_file_hash = NULL
)
}
\arguments{
\item{chain}{List of method steps}

\item{dependencies}{Vector of node IDs this depends on}

\item{input_file_hash}{Optional file hash for root nodes}
}
\value{
Node definition
}
\description{
Create a pipeline node definition
}
