% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_files.R
\name{upload_file}
\alias{upload_file}
\title{Upload content to the HPC resource}
\usage{
upload_file(config, content, filename)
}
\arguments{
\item{config}{API configuration created by create_api_config}

\item{content}{Content to upload (raw vector, character, or other object)}

\item{filename}{Name to use for the uploaded content}
}
\value{
TRUE if the content was successfully uploaded or already exists
}
\description{
Upload content to the HPC resource
}
\examples{
\dontrun{
config <- create_api_config("http://localhost", 9000, "please_change_me")
content <- "Hello, World!"
upload_file(config, content, "hello.txt")
}
}
